+++ cuenca/__init__.py
@@ -12,11 +12,13 @@ __all__ = [
     'Commission',
     'Deposit',
     'LoginToken',
+    'Saving',
     'ServiceProvider',
     'Statement',
     'Transfer',
     'UserCredential',
     'UserLogin',
+    'WalletTransaction',
     'WhatsappTransfer',
     'configure',
     'get_balance',
@@ -38,11 +40,13 @@ from .resources import (
     Commission,
     Deposit,
     LoginToken,
+    Saving,
     ServiceProvider,
     Statement,
     Transfer,
     UserCredential,
     UserLogin,
+    WalletTransaction,
     WhatsappTransfer,
 )
 from .version import __version__
