+++ requirements.txt
@@ -1,3 +1,3 @@
-requests==2.26.0
-cuenca-validations==0.9.10
+requests==2.25.1
+cuenca-validations==0.9.11.dev5
 dataclasses>=0.7;python_version<"3.7"
