+++ tests/resources/cassettes/test_saving_create.yaml
@@ -1,0 +1,60 @@
+interactions:
+- request:
+    body: '{"name": "my new car", "category": "vehicle", "goal_amount": 100000, "goal_date":
+      "2022-08-13T00:59:27.248084"}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Authorization:
+      - DUMMY
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '111'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - cuenca-python/0.7.10
+      X-Cuenca-Api-Version:
+      - '2020-03-19'
+      X-Cuenca-LoginId:
+      - ULIfKcTXnXRv-UV8HPObyxYw
+    method: POST
+    uri: https://sandbox.cuenca.com/savings
+  response:
+    body:
+      string: '{"id":"lA224Vz0MJR6C6C27gvalk_Q","created_at":"2021-08-13T00:59:28.069546","updated_at":"2021-08-13T00:59:28.069551","user_id":"US3zGWi1n852bTyqD1wCZ1ft","balance":0,"deactivated_at":null,"name":"my
+        new car","category":"vehicle","goal_amount":100000,"goal_date":"2022-08-13T00:59:27.248084"}'
+    headers:
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '291'
+      Content-Type:
+      - application/json
+      Date:
+      - Fri, 13 Aug 2021 00:59:28 GMT
+      X-Amzn-Trace-Id:
+      - Root=1-6115c3ef-3cffd81045e9540a5b1a90c3;Sampled=0
+      X-Request-Time:
+      - 'value: 0.627'
+      x-amz-apigw-id:
+      - D-uNfEA-CYcFbRQ=
+      x-amzn-Remapped-Connection:
+      - keep-alive
+      x-amzn-Remapped-Content-Length:
+      - '291'
+      x-amzn-Remapped-Date:
+      - Fri, 13 Aug 2021 00:59:28 GMT
+      x-amzn-Remapped-Server:
+      - nginx/1.20.1
+      x-amzn-Remapped-x-amzn-RequestId:
+      - 0d84b80d-24d0-49a0-9a34-94cebf94c1f8
+      x-amzn-RequestId:
+      - 3a975ded-4382-4a8c-9c61-3ad956762d9e
+    status:
+      code: 201
+      message: Created
+version: 1
