+++ tests/resources/cassettes/test_saving_retrieve.yaml
@@ -1,0 +1,55 @@
+interactions:
+- request:
+    body: null
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Authorization:
+      - DUMMY
+      Connection:
+      - keep-alive
+      User-Agent:
+      - cuenca-python/0.7.10
+      X-Cuenca-Api-Version:
+      - '2020-03-19'
+      X-Cuenca-LoginId:
+      - ULIfKcTXnXRv-UV8HPObyxYw
+    method: GET
+    uri: https://sandbox.cuenca.com/savings/lAob5rOC0jSj6UC5RAbwnSnA
+  response:
+    body:
+      string: '{"id":"lAob5rOC0jSj6UC5RAbwnSnA","created_at":"2021-08-13T00:36:55.471000","updated_at":"2021-08-13T00:36:55.471000","user_id":"US3zGWi1n852bTyqD1wCZ1ft","balance":0,"deactivated_at":"2021-08-13T00:50:40.133000","name":"my
+        new home","category":"property","goal_amount":200000,"goal_date":"2021-08-23T00:55:27.870000"}'
+    headers:
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '317'
+      Content-Type:
+      - application/json
+      Date:
+      - Fri, 13 Aug 2021 00:59:29 GMT
+      X-Amzn-Trace-Id:
+      - Root=1-6115c3f1-71aac8b85b6ca0fb61f4ec0d;Sampled=0
+      X-Request-Time:
+      - 'value: 0.289'
+      x-amz-apigw-id:
+      - D-uNsFT1CYcFnqw=
+      x-amzn-Remapped-Connection:
+      - keep-alive
+      x-amzn-Remapped-Content-Length:
+      - '317'
+      x-amzn-Remapped-Date:
+      - Fri, 13 Aug 2021 00:59:29 GMT
+      x-amzn-Remapped-Server:
+      - nginx/1.20.1
+      x-amzn-Remapped-x-amzn-RequestId:
+      - ac85db89-f1c8-4693-96a4-a0681971aa9a
+      x-amzn-RequestId:
+      - 6d415fa6-5325-4075-b503-31b7cbf09eaf
+    status:
+      code: 200
+      message: OK
+version: 1
