+++ tests/resources/cassettes/test_saving_update.yaml
@@ -1,0 +1,59 @@
+interactions:
+- request:
+    body: '{"name": "my new home", "category": "property", "goal_amount": 200000}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Authorization:
+      - DUMMY
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '70'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - cuenca-python/0.7.10
+      X-Cuenca-Api-Version:
+      - '2020-03-19'
+      X-Cuenca-LoginId:
+      - ULIfKcTXnXRv-UV8HPObyxYw
+    method: PATCH
+    uri: https://sandbox.cuenca.com/savings/lAob5rOC0jSj6UC5RAbwnSnA
+  response:
+    body:
+      string: '{"id":"lAob5rOC0jSj6UC5RAbwnSnA","created_at":"2021-08-13T00:36:55.471000","updated_at":"2021-08-13T00:36:55.471000","user_id":"US3zGWi1n852bTyqD1wCZ1ft","balance":0,"deactivated_at":"2021-08-13T00:50:40.133000","name":"my
+        new home","category":"property","goal_amount":200000,"goal_date":"2021-08-23T00:55:27.870000"}'
+    headers:
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '317'
+      Content-Type:
+      - application/json
+      Date:
+      - Fri, 13 Aug 2021 00:59:30 GMT
+      X-Amzn-Trace-Id:
+      - Root=1-6115c3f2-5657920a7f155d406bd87c22;Sampled=0
+      X-Request-Time:
+      - 'value: 0.317'
+      x-amz-apigw-id:
+      - D-uN3FR4iYcFrfA=
+      x-amzn-Remapped-Connection:
+      - keep-alive
+      x-amzn-Remapped-Content-Length:
+      - '317'
+      x-amzn-Remapped-Date:
+      - Fri, 13 Aug 2021 00:59:30 GMT
+      x-amzn-Remapped-Server:
+      - nginx/1.20.1
+      x-amzn-Remapped-x-amzn-RequestId:
+      - c9393e15-a1c0-4cac-a408-6fae24d5e1ff
+      x-amzn-RequestId:
+      - 9e19a455-ad7a-4290-87ac-9e2e0064abd8
+    status:
+      code: 200
+      message: OK
+version: 1
